Tabela de estatísticas de arremessadores

Estatísticas de arremessadores

Format

Um data frame com 46.699 linhas e 30 colunas

id_jogador

ID do jogador

id_ano

Ano

ordem_equipes

Ordem na qual o jogador se mudou entre equipes na mesma temporada

id_equipe

ID da equipe (fator)

id_liga

ID da liga (fator com os níveis AA, AL, FL, NL, PL, UA)

jogos_ganhados

Jogos disputados ganhados

jogos_perdidos

Jogos disputados perdidos

jogos_disputados

Jogos disputados

jogos_iniciados

Jogos disputados iniciados

jogos_completos

Número de jogos completos que o jogador lançou

shutouts

Número de shutouts (jogos completos sem permitir corridas) que o jogador lançou

jogos_salvos

Jogos salvos

IPouts

Outs lançadas à equipe adversária (igual a entradas lançadas x 3)

hits

Hits cedidos ao oponente

corridas_merecidas

Corridas merecidas (por culpa do arremessador) cedidas

homerun

Número de homeruns

BB

Base por bolas cedidas ao oponente

strikeout

Strikeout dados ao oponente

media_rebatida_oponente

Média de rebatidas do oponente

media_corridas_merecidas

Média de corridas merecidas cedidas (normalizada para 9 entradas lançadas)

IBB

Base por bolas intencionais cedidas ao oponente

lances_desviados

Wild pitches (arremessos ruins que permitem o avanço de corredores adversários) lançados pelo arremessador

HBP

Número de avanços cedidos à primeira base por atingir o rebatedor com o arremesso

BK

Balks (movimento ilegal do corpo pelo lançador)

BFP

Rebatedores que o arremessador enfrentou

jogos_terminados

Jogos em que o lançador termina o jogo

corridas

Corridas cedidas (merecidas e nao merecidas)

toques_sacrificio

Toques de sacrifício (impulsiona o avanço de um corredor) feitos pelo (rebatedor) oponente

fly_sacrificio

Flys (rebatidas altas no campo externo) de sacrifício (impulsiona uma corrida) contra o arremessador

eliminacao_dupla

Arremessos que levam a uma rebatida rasteira com eliminação dupla na mesma jogada

See also