Dados de habitação em Ames

Base dados descrita em De Cock (2011) com 82 características de 2930 propriedades vendidas em Ames, Iowa, EUA.

casas

Format

Um data.frame com 2930 linhas e 82 colunas.

ordem

Número da observação (id)

pid

Código de identificação análogo ao CEP no Brasil (Parcel identification number)

moradia_classe

Tipo de moradia envolvida na venda

moradia_zoneamento

Classificação do zoneamento da venda

lote_fachada

Porção da fachada cercada por uma rua (pés)

lote_area

Área do lote (pés quadrados)

rua_tipo

Tipo de rua que dá acesso à propriedade

beco_tipo

Tipo do beco que dá acesso à propriedade

lote_formato

Formato da propriedade

terreno_contorno

Nivelamento da propriedade. Valores: bancada (aumento rápido e significativo do nível da rua para o prédio), encosta (declive significativo de lado a lado), depressão, aproximadamente plana/nivelada.

utilidades

Tipo de utilizades disponíveis (eletricidade, gás, água e esgoto)

lote_config

Configuração do lote

terreno_declive

Declive do terreno

vizinhanca

Localizações físicas dentro dos limites da cidade de ames

condicao_1

Proximidade da propriedade estradas ou rodovias importantes

condicao_2

Proximidade da propriedade de estradas ou rodovias importantes (caso uma segunda exista)

moradia_tipo

Tipo de moradia. Valores: família única, conversão para 2 famílias (originalmente construída para 2 famílias), duplex, unidade final de moradia (Townhouse End Unit), unidade interna de moradia (Townhouse Inside Unit).

moradia_estilo

Estilo da moradia

geral_qualidade

Avaliação do material e acabamento geral da casa

geral_condicao

Avaliação da condição geral da casa

construcao_ano

Ano de construção da casa

remodelacao_ano

Ano de remodelação da casa

telhado_estilo

Estilho do telhado

telhado_material

Material do qual o telhado é feito

exterior_cobertura_1

Cobertura do exterior da casa

exterior_cobertura_2

Cobertura do exterior da casa (se tiver mais de uma)

alvenaria_tipo

Tipo de alvenaria da casa

alvenaria_area

Área da alvenaria (pés quadrados)

exterior_qualidade

Avaliação da qualidade do material usado no exterior da casa

exterior_condicao

Avaliação da condição do exterior da casa

fundacao_tipo

Tipo da fundação da casa

porao_qualidade

Avaliação da altura do porão

porao_condicao

Avaliação da condição do porão

porao_exposicao

Refere-se a saída ou paredes no nível do jardim

porao_acabamento_1

Avaliação da área do porão com acabamento

porao_area_com_acabamento_1

Área do porão com acabamento (em pés quadrados)

porao_acabamento_2

Avaliação da área do porão com acabamento (se houver várias)

porao_area_com_acabamento_2

Área do porão com acabamento (em pés quadrados)

porao_area_sem_acabamento

Área do porão sem acabamento (em pés quadrados)

porao_area_total

Área total do porão (em pés quadrados)

porao_num_banheiros

Número de banheiros do porão

porao_num_banheiros_lavabos

Número de lavabos do porão

aquecimento_tipo

Tipo de aquecimento instalado na casa

aquecimento_qualidade_condicao

Avaliação da qualidade e condição do aquecimento

ar_condicionado_central

Ar condicionado central da casa

sistema_eletrico_tipo

Tipo de sistema elétrico da casa

primeiro_andar_area

Área do primeiro andar (pés quadrados)

segundo_andar_area

Área do segundo andar (pés quadrados)

acabamento_baixa_qualidade_area

Área da casa com acabamento de baixa qualidade (pés quadrados)

acima_solo_area

Área da casa acima do solo (pés quadrados)

acima_solo_num_banheiros

Número de banheiros acima do solo

acima_solo_num_lavabos

Número de lavabos acima do solo

acima_solo_num_quartos

Número de quartos da casa

acima_solo_num_cozinhas

Número de cozinhas da casa

cozinha_qualidade

Qualidade da cozinha

acima_solo_num_comodos

Número de cômodos acima do solo

funcional

Funcionalidade da casa (indica o quanto a casa está apta para moradia)

total_num_lareiras

Número total de lareiras

lareira_qualidade

Qualidade da(s) lareira(s). Valores: excelente (fogueira de alvenaria exepecional), boa (lareira de alvenaria no nível principal), regular (fogueira pré-fabricada no porão), média (lareira pré-fabricada na sala de estar principal ou lareira de alvenaria no porão), ruim (fogão Ben Franklin), sem lareira.

garagem_tipo

Localização da garagem. Valores: mais do que 1 tipo de garagem, anexada à casa, garagem no porão, embutida (parte da casa - tipicamente tem cômodo acima da garagem), abrigo com teto construído ao lado da casa (carport), garagem não anexada à casa, sem garagem.

garagem_ano_construcao

Ano de construção da garagem

garagem_acabamento

Tipo de acabamento da garagem

garagem_capacidade_carros

Capacidade de carros da garagem

garagem_area

Área da garagem (pés quadrados)

garagem_qualidade

Qualidade da garagem

garagem_condicao

Condição da garagem

entrada_veiculo_pavimentada

Pavimentação da entrada de veículos

deck_madeira_area

Área do deck de madeira da casa (em pés quadradas)

varanda_aberta_area

Área da varanda coberta da casa (em pés quadradas)

varanda_fechada_area

Área da varanda fechada da casa (em pés quadradas)

varanda_3ssn_area

Área da varanda 'three season' da casa (em pés quadradas)

varanda_com_tela_area

Área da varanda com tela da casa (em pés quadradas)

piscina_area

Área da piscina (em pés quadradas)

piscina_qualidade

Qualidade da piscina

cerca_qualidade

Qualidade da cerca

funcionalidades_diversas

Funcionalidades diversas da casa (não incluidas nas outras categorias). Valores: elevador, segunda garagem (se não descrito na seção garagem), galpão, quadra de tênis, outro, nenhuma.

funcionalidades_valor

Valor das funcionalidades diversas

venda_mes

Mês da venda

venda_ano

Ano da venda

venda_tipo

Tipo da venda

venda_condicao

Condição da venda. Valores: venda anormal (troca, execução hipotecária, venda a descoberto), compra de terreno adjacente, alocação (duas propriedades vinculadas com escrituras separadas, normalmente em condomínio com garagem), venda entre membros de família, venda normal, casa não foi concluída na última avaliação (associada a novas casas)

venda_valor

Valor da venda

Source

http://jse.amstat.org/v19n3/decock/DataDocumentation.txt

See also