Vetores de caracteres do pacote stringr

Amostra de vetores de caracteres para praticar a manipulação de strings

frases

frutas

palavras

Format

Um vetor

An object of class character of length 37.

An object of class character of length 116.

An object of class character of length 267.

Details

frases - é uma lista de frases escritas em português

frutas - é uma seleção e adaptação ao português dos dados de prática contidos em stringr::fruit.

palavras - é uma lista de palavras frequentes usadas em português brasileiro

Examples

length(frases)
#> [1] 37
frases[1:5]
#> [1] "Ao verme que primeiro roeu as frias carnes do meu cadáver dedico como saudosa lembrança estas Memórias Póstumas."
#> [2] "Não tive filhos, não transmiti a nenhuma criatura o legado da nossa miséria."                                    
#> [3] "Ao vencedor, as batatas."                                                                                        
#> [4] "Marcela amou-me durante quinze meses e onze contos de réis."                                                     
#> [5] "Aos quarenta e cinco anos estava ainda moço, e para a minha cura não havia outro remédio senão um enterro."      

length(frutas)
#> [1] 116
frutas[1:5]
#> [1] "abacate" "abacaxi" "açaí"    "acerola" "ameixa" 

#' length(palavras)
palavras[1:5]
#> [1] "ser"   "dizer" "ter"   "ir"    "estar"