Dados de economia de combustível de 1999 a 2008 para 38 modelos populares de carros

Este conjunto de dados contém um subconjunto dos dados de economia de combustível que a EPA disponibiliza em http://fueleconomy.gov. Ele contém apenas modelos que tiveram um novo lançamento todos os anos entre 1999 e 2008 - isso foi usado como um proxy para a popularidade do carro

milhas

Format

Um data.frame com 234 linhas e 11 colunas

fabricante

Nome do fabricante

modelo

Nome do modelo

cilindrada

Capacidade do motor, em litros

ano

Ano de fabricação

cilindros

Número de cilindros

transmissao

Tipo de transmissão

tracao

Tipo de tração (d = dianteira, t = traseira, 4 = 4x4)

cidade

Milhas da cidade por galão

rodovia

Milhas rodoviárias por galão

combustivel

Tipo de combustível (p = premium, r = regular, e = etanol, d = diesel, g = gás natural comprimido)

classe

tipo de carro

See also