Tabela de pessoas

Nomes de jogadores, data de nascimento e informações bibliográficas. Este arquivo é usado para obter detalhes sobre os jogadores listados em rebatedores, arremessadores e outros arquivos nos quais os jogadores são identificados por id_jogador

Format

Um data frame com 19.617 linhas e 26 colunas

id_jogador

ID do jogador

ano_nascimento

Ano de nascimento

mes_nascimento

Mês de nascimento

dia_nascimento

Dia de nascimento

pais_nascimento

País de nascimento

estado_nascimento

Estado de nascimento

cidade_nascimento

Cidade de nascimento

ano_falecimento

Ano de falecimento

mes_falecimento

Mês de falecimento

dia_falecimento

Dia de falecimento

pais_falecimento

País de falecimento

estado_falecimento

Estado de falecimento

cidade_falecimento

Cidade de falecimento

nome

Nome

sobrenome

Sobrenome

nome_proprio

Nome próprio

peso

Peso em libras

altura

Altura em polegadas

rebatedor

Se, como rebatedor, você é destro (D) ou canhoto (C)

arremessador

Se, como arremessador, você é destro (D) ou canhoto (C)

data_primeiro_jogo

Data do primeiro jogo na primeira divisão

data_ultimo_jogo

Data do último jogo na primeira divisão

id_retrosheet

Código do jogador usado em Retrosheet (http://www.retrosheet.org/)

id_brw

Código do jogador usado em Baseball Reference website (http://www.baseball-reference.com/)

data_falecimento

Data de falecimento em formato yyyy-mm-dd

data_nascimento

Data de nascimento em formato yyyy-mm-dd

See also