Tabela de estatíticas de Beisebol

Estatísticas de rebatedores

Format

Um data frame com 105.861 linhas e 22 colunas

id_jogador

ID do jogador

id_ano

Ano

ordem_equipes

Ordem na qual o jogador se mudou entre equipes na mesma temporada

id_equipe

ID da equipe (fator)

id_liga

ID da liga (fator com os níveis AA, AL, FL, NL, PL, UA)

jogos

Número de jogos em que o jogador participou oficialmente

vez_bastao

Número de vezes ao bastão válidas

corridas

Corridas anotadas

rebatidas_simples

Número de vezes em que o jogador chegou à primeira base com uma rebatida válida

rebatidas_duplas

Número de vezes em que o jogador chegou à segunda base com uma rebatida válida

rebatidas_triplas

Número de vezes em que o jogador chegou à terceira base com uma rebatida válida

homerun

Número de homeruns

corridas_impulsionadas

Corridas impulsionadas como resultado da vez ao bastão do rebatedor

bases_roubadas

Bases roubadas

capturado_roubando

Eliminações em tentativa de roubo de base

base_bolas

Base por bolas

strikeouts

Eliminações por strikes

base_intencional

Base por bolas intencional

HBP

Vezes atingidas pelo arremessador (Hit by pitch - HBP)

toque_sacrificio

Toque de sacrifício

fly_sacrificio

Flys (rebatidas altas no campo externo) de sacrifício (impulsiona uma corrida)

eliminacao_dupla_induzida

Rebatidas rasteiras induzindo a uma eliminação dupla

See also