Dados de voos

Dados temporais de todos os voos que partiram de Nova York (JFK, LGA e EWR) em 2013

voos

Format

Um data.frame com 336.776 linhas e 19 colunas

ano

Ano da data de saída do voo

mes

Mês da data de saída do voo

dia

Dia da data de saída do voo

horario_saida

Horário real de saída do voo (formato HHMM ou HMM), hora local

saida_programada

Horário programado para a saída do voo (formato HHMM ou HMM), hora local

atraso_saida

Atraso de saída em minutos. Valores negativos indicam saída adiantada

horario_chegada

Horário real de chegada do voo (formato HHMM ou HMM), hora local

chegada_prevista

Horário programado para a chegada do voo (formato HHMM ou HMM), hora local

atraso_chegada

Atraso de chegada em minutos. Valores negativos indicam chegada adiantada

companhia_aerea

Abreviação de dois dígitos da companhia aérea. Ver `companhias_aereas` para obter o nome da companhia aérea

voo

Número do voo

codigo_cauda

Código da cauda do avião (empenagem)

origem

Origem do voo. Ver `aeroportos` para metadados adicionais

destino

Destino do voo. Ver `aeroportos` para metadados adicionais

tempo_voo

Quantidade de tempo no ar, em minutos

distancia

Distância entre aeroportos, em milhas

hora

Hora do horário programado para saída do voo

minuto

Minutos do horário programado para a saída do voo

data_hora

Data e hora programados para o voo em formato POSIXct. Juntamente com a coluna origem, pode ser usado para juntar dados de voos aos dados meteorológicos disponíveis em `dados_clima`

See also